+++ chaff-reader/src/main/scala/x7c1/chaff/reader/Reader.scala
@@ -1,5 +1,6 @@
 package x7c1.chaff.reader
 
+
 trait Reader[X, A] extends BaseReader[X, A] {
 
   override type Type[T] = Reader[X, T]
